+++ IGC/Compiler/CISACodeGen/TypeDemote.cpp
@@ -293,7 +293,7 @@ bool TypeDemote::demoteOnBasicBlock(BasicBlock* BB) const {
             if (CI && (CI->getOpcode() == Instruction::ZExt ||
                 CI->getOpcode() == Instruction::SExt)) {
                 unsigned VS = EEI->getVectorOperandType()->getVectorNumElements();
-                unsigned N = CI->getSrcTy()->getPrimitiveSizeInBits();
+                unsigned N = (unsigned int)CI->getSrcTy()->getPrimitiveSizeInBits();
                 unsigned Bound = (N < 32) ? (1U << N) : UINT32_MAX;
                 if (VS <= Bound) {
                     EEI->setOperand(1, CI->getOperand(0));
